place extract_custom_headers function in a common file
