+++ instana/instrumentation/flask/with_blinker.py
@@ -11,26 +11,12 @@ import opentracing.ext.tags as ext
 from ...log import logger
 from ...util.secrets import strip_secrets_from_query
 from ...singletons import agent, tracer
+from common import extract_custom_headers
 
 import flask
 from flask import request_started, request_finished, got_request_exception
 
 path_tpl_re = re.compile('<.*>')
-
-
-def extract_custom_headers(span, headers, format):
-    if agent.options.extra_http_headers is None:
-        return
-    try:
-        for custom_header in agent.options.extra_http_headers:
-            # Headers are available in this format: HTTP_X_CAPTURE_THIS
-            flask_header = ('HTTP_' + custom_header.upper()).replace('-', '_') if format else custom_header
-
-            if flask_header in headers:
-                span.set_tag("http.header.%s" % custom_header, headers[flask_header])
-
-    except Exception:
-        logger.debug("extract_custom_headers: ", exc_info=True)
 
 
 def request_started_with_instana(sender, **extra):
